Add ST7586S LCD emulation, OS16 boot system, build infrastructure, and toolchain updates
em16: Replace UC1608 LCD with ST7586S controller emulation, add VIA 6522,
ADB controller, L28 MCU, and Video subsystems.
OS16: Add boot loader, module system, FAT12 disk support, terminal app.
as: Refactor code generation, fix stack-relative addressing.
oc: Update Oberon compiler code generation and test suite.

+++ oc/include/codegen.h
@@ -29,8 +29,8 @@ extern bool longi;
 
 #define maxReloc 1024
 
-extern short int relocC;
-extern short int reloc[maxReloc];
+extern int relocC;
+extern int reloc[maxReloc];
 
 void codegen_gen(OpCode opcode, AddrMode mode, int value1, int value2);
 void codegen_byte(int b);
